requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langgraph
langchain
langchain-openai
pydantic
python-dotenv==1.0.0
jinja2==3.1.3
aiofiles==23.2.1
